requirements.txt
# Dependencies for detector
torch>=1.8.0
torchvision>=0.9.0
pillow>=8.0.0
pycaw>=20230407
comtypes>=1.2.0
selenium>=4.0.0
webdriver-manager>=4.0.0

# Training dependencies (optional - only needed for train_nfl_classifier.py)
scikit-learn>=0.24.0
numpy>=1.20.0